Import node fix:
Replacing the mention "user" for the currenet user importing wasn't working.
The metadata_wad column was removed so as to put the temporary table in memory, rather than on disk because of the TEXT/BLOB mention.
It's back with a size of 2048 characters, so far our use has been lower than 1024 characters, it 'should' be fine for now.

Changing data post copy would slow down to some extent importing since we want to change data that is on disk, so this solution is put aside for now.