Query fix:
The basic submit query would search for a user in a specified group if it exists before executing any changes.
But specific right to a specific user are put in a group named after a user without him being inside.
Now we add the user inside.

Also, for an existing rights, the MySQL query would just copy the existing value rather than remove them and only let the reading rights.