Query fix:
When resetting rights, the query is using an old table name.
Also something with the way Oracle wanted his queries but I don't remember,
the second query was using the same variable than the first to find what is the current group rights, so when we actually do need to set variables to change rights, it wouldn't work.