Node import fix:
When importing a node that has a 'user' directive, the first import need to create a specific user group for the portfolio (done correctly) and one for the temporary table where we manipulate data first. But that last command wasn't sent so it wouldn't work when we first add a new object, and works the second time.

Also checking for a null pointer when recontructing the portfolio