Fixing public access:
Checking if the portfolio is public previously relied on checking if a node was in a portfolio from the rights table.
Obviously this doesn't work on a referential where rights aren't created.

Next would be the evaluation order, we would check if a user has right on it, which he doesn't since it's not even instanciated, then bypass the check if it's public or not.